Added a feature and fixed some bugs
So, the image processing did not work right so I fixed it, it took me quite an awful amount of time to do that because I couldn't focus for more than 5 minutes but it's done. Also, I added a feature that outputs a file that you can use in a datapack in minecraft to automate the process even further.