Added comments and small improvements
Added comments on every file for what it's input is, out, and purpose. I also changed how the colour detection works and what I mean by that is I made it that if it finds a colour that is not included in the config file it will just throw an exception to the user informing the to update the file. After that I did more changes but  they are not noteworthy.